src/requirements.txt
selenium==4.11.2
black==23.7.0
pandas==2.0.3
lxml==4.9.3
html5lib==1.1
beautifulsoup4==4.12.2
undetected_chromedriver==3.5.3
dadata==21.10.1
psycopg2-binary==2.9.7
pre-commit==3.3.3
